@@ -1,12 +1,12 @@
 name: libphonenumber
 description: This is a Flutter implementation of libphonenumber. It includes only a few features at the moment, and are added when more functionality is desired.
-version: 1.0.2
+version: 2.0.0
 homepage: https://github.com/emostar/flutter-libphonenumber
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.2.3
+  meta: ^1.3.0
 
 flutter:
   plugin:
@@ -18,5 +18,5 @@ flutter:
         pluginClass: LibphonenumberPlugin
 
 environment:
-  sdk: ">=1.19.0 <3.0.0"
-  flutter: ">=1.12.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
