@@ -1,12 +1,11 @@
 import 'dart:async';
 
 import 'package:flutter/services.dart';
-import 'package:meta/meta.dart';
 
 class RegionInfo {
-  String regionPrefix;
-  String isoCode;
-  String formattedPhoneNumber;
+  String? regionPrefix;
+  String? isoCode;
+  String? formattedPhoneNumber;
 
   RegionInfo({this.regionPrefix, this.isoCode, this.formattedPhoneNumber});
 
@@ -34,9 +33,9 @@ enum PhoneNumberType {
 class PhoneNumberUtil {
   static const MethodChannel _channel = const MethodChannel('codeheadlabs.com/libphonenumber');
 
-  static Future<bool> isValidPhoneNumber({
-    @required String phoneNumber,
-    @required String isoCode,
+  static Future<bool?> isValidPhoneNumber({
+    required String phoneNumber,
+    required String isoCode,
   }) async {
     try {
       return await _channel.invokeMethod('isValidPhoneNumber', {
@@ -49,9 +48,9 @@ class PhoneNumberUtil {
     }
   }
 
-  static Future<String> getNameForNumber({
-    @required String phoneNumber,
-    @required String isoCode,
+  static Future<String?> getNameForNumber({
+    required String phoneNumber,
+    required String isoCode,
   }) async {
     return await _channel.invokeMethod('getNameForNumber', {
       'phone_number': phoneNumber,
@@ -59,9 +58,9 @@ class PhoneNumberUtil {
     });
   }
 
-  static Future<String> normalizePhoneNumber({
-      @required String phoneNumber,
-      @required String isoCode,
+  static Future<String?> normalizePhoneNumber({
+      required String phoneNumber,
+      required String isoCode,
     }) async {
       return await _channel.invokeMethod('normalizePhoneNumber', {
         'phone_number': phoneNumber,
@@ -70,13 +69,13 @@ class PhoneNumberUtil {
     }
 
   static Future<RegionInfo> getRegionInfo({
-    @required String phoneNumber,
-    @required String isoCode,
+    required String phoneNumber,
+    required String isoCode,
   }) async {
-    Map<dynamic, dynamic> result = await _channel.invokeMethod('getRegionInfo', {
+    Map<dynamic, dynamic> result = await (_channel.invokeMethod('getRegionInfo', {
       'phone_number': phoneNumber,
       'iso_code': isoCode,
-    });
+    }) as FutureOr<Map<dynamic, dynamic>>);
 
     return RegionInfo(
       regionPrefix: result['regionCode'],
@@ -86,22 +85,22 @@ class PhoneNumberUtil {
   }
 
   static Future<PhoneNumberType> getNumberType({
-    @required String phoneNumber,
-    @required String isoCode,
+    required String phoneNumber,
+    required String isoCode,
   }) async {
-    int result = await _channel.invokeMethod('getNumberType', {
+    int? result = await _channel.invokeMethod('getNumberType', {
       'phone_number': phoneNumber,
       'iso_code': isoCode,
     });
     if (result == -1) {
       return PhoneNumberType.unknown;
     }
-    return PhoneNumberType.values[result];    
+    return PhoneNumberType.values[result!];    
   }
   
-  static Future<String> formatAsYouType({
-    @required String phoneNumber,
-    @required String isoCode,
+  static Future<String?> formatAsYouType({
+    required String phoneNumber,
+    required String isoCode,
   }) async {
     return await _channel.invokeMethod('formatAsYouType', {
       'phone_number': phoneNumber,
